pyproject.toml
[tool.poetry]
name = "langchain-test"
version = "0.1.0"
description = "langchain tests"
authors = ["Sebastian Montero <sebastian.montero@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.1"
openai = "^0.27.6"
pinecone-text = "0.1.1"
tiktoken = "^0.3.3"
pandas = "^2.0.1"
evaluate = "^0.4.0"
langchain = {path = "../langchain"}
langchain-util = {path = "../langchain-util", develop = true}
apify-client = "^1.1.1"
firebase-admin = "^6.1.0"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
marko = "^1.3.0"
apollo-client = "^2.1.2"
gql = "^3.4.1"
requests-toolbelt = "^1.0.0"
pypdf2 = "^3.0.1"
google-api-python-client = "^2.87.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
nltk = "^3.8.1"
python-docx = "^0.8.11"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
